Redo the override by swizzling on DVTLayoutManager instead of subclassing it. Also clean up a lot of plugin boilerplate.
Playgrounds use a DVTLayoutManager subclass and the old code caused a crash because it substituted its own subclass instead, which didn't have the same methods. By swizzling DVTLayoutManager, the new code gets into playgrounds too, but it doesn't hit other NSLayoutManagers. It could subclass the playground layout manager instead, but this is more straightforward and achieves the same thing.

All the notification shenanigans that led up to the dynamic subclassing appear to be unnecessary and have been simplified down to performing the swizzling from +pluginDidLoad:.